Do some cleanup: - Remove boilerplate code - Separate timeline item from the timeline - Separate styles - Remove useless styles - Remove useless assets - Separate hooks to make them more readable
