+++ src/components/timeline-item.css
@@ -1,0 +1,118 @@
+.timeline-item {
+    padding: 10px 20px;
+    margin: 20px 0;
+    background: #f9f9f9;
+    border-radius: 8px;
+    position: relative;
+    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+    width: 90%;
+    clear: both;
+    transition: background-color 0.3s ease;
+}
+
+.timeline-content > h3 {
+    color: #c10101;
+    font-size: 1.5em;
+    margin: 0;
+}
+
+.timeline-separator {
+    content: '';
+    width: 6px;
+    background-color: #009688;
+    justify-self: center;
+    height: 30px;
+    border-radius: 4px;
+}
+
+.timeline-item.active.left {
+    background-color: #e0f7fa;
+    border-right: 4px solid #009688;
+}
+
+.timeline-item.active.right {
+    background-color: #e0f7fa;
+    border-left: 4px solid #009688;
+}
+
+.timeline-item.left::before,
+.timeline-item.right::before {
+    content: '';
+    position: absolute;
+    top: 15px;
+    width: 24px;
+    height: 24px;
+    background: #f9f9f9;
+    border: 4px solid #009688;
+    border-radius: 50%;
+}
+
+.timeline-item.left::before {
+    right: -12px;
+}
+
+.timeline-item.right::before {
+    left: -12px;
+}
+
+.timeline-content {
+    padding: 20px;
+    background: white;
+    border-radius: 8px;
+    position: relative;
+}
+
+.timeline-content h2 {
+    margin: 0;
+    color: #009688;
+}
+
+.timeline-content p {
+    margin: 8px 0;
+}
+
+
+
+/* Medium screens (tablets, 768px and up) */
+@media (min-width: 768px) {
+    .timeline-item {
+        width: 45%;
+    }
+
+    .timeline-item.left {
+        justify-self: flex-start;
+        clear: both;
+        margin-left: -12px;
+    }
+
+    .timeline-item.right {
+        justify-self: flex-end;
+        clear: both;
+        margin-right: -12px;
+    }
+}
+
+/* Small screens (phones, up to 767px) */
+@media (max-width: 767px) {
+    .timeline-item {
+        width: 90%;
+        float: none;
+        clear: both;
+        transform: translateX(0);
+        margin-left: auto;
+        margin-right: auto;
+    }
+
+    .timeline-item.left::before,
+    .timeline-item.right::before {
+        left: 50%;
+        transform: translateX(-50%);
+    }
+
+    .timeline-item.left.active,
+    .timeline-item.right.active {
+        background-color: #e0f7fa;
+        border-left: 4px solid #009688;
+        border-right: 4px solid #009688;
+    }
+}
