+++ src/utils/debounce.ts
@@ -1,0 +1,7 @@
+export const debounce = (func: () => void, delay: number) => {
+    let timeout: number;
+    return () => {
+        clearTimeout(timeout);
+        timeout = window.setTimeout(func, delay);
+    };
+}
